+++ cli/src/commands/tag/delete.rs
@@ -50,7 +50,12 @@ pub fn cmd_tag_delete(
 ) -> Result<(), CommandError> {
     let mut workspace_command = command.workspace_helper(ui)?;
     let repo = workspace_command.repo().clone();
-    let matched_tags = find_local_tags(repo.view(), &args.names)?;
+    let matched_tags = find_local_tags(ui, repo.view(), &args.names)?;
+    if matched_tags.is_empty() {
+        writeln!(ui.status(), "No tags to delete.")?;
+        return Ok(());
+    }
+
     let mut tx = workspace_command.start_transaction();
     for (name, _) in &matched_tags {
         tx.repo_mut()